refactor(tool-shell): implement single responsibility pattern
- Split monolithic main.go (388 lines) into focused modules:
  - main.go: Clean entry point with plugin metadata (60 lines)
  - shell_tool.go: Execute method and command routing
  - shell_setup.go: Shell configuration setup and initialization
  - shell_switcher.go: Shell switching with validation and safety
  - shell_config.go: Configuration status and management
  - shell_backup.go: Timestamped configuration backups

- Enhanced functionality:
  - Improved shell detection and validation
  - Better configuration status reporting with emojis
  - Enhanced backup system with timestamped directories
  - Comprehensive shell support (bash/zsh/fish)
  - Better error handling and user guidance

- Follows same modular pattern as package-manager-apt
- Each file maintains single responsibility
- Improved maintainability and testability
